projects/penicillin/requirements.txt
Flask==2.3.2
flask_injector==0.15.0
flask_httpauth==4.8.0
injector==0.21.0
gunicorn==21.2.0
azure-monitor-opentelemetry==1.0.0b13
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-flask==0.38b0
azure-identity==1.13.0
azure-data-tables==12.4.3